apps/worker/pyproject.toml
[project]
name = "photoprune-worker"
version = "0.0.0"
description = "PhotoPrune Celery worker skeleton"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.4.0",
    "redis>=5.0.8",
    "python-dotenv>=1.0.1"
]

[dependency-groups]
dev = [
    "ruff==0.6.8",
    "black==24.8.0",
    "mypy==1.11.2",
    "pytest==8.3.2",
    "pytest-cov==5.0.0",
    "pip-audit==2.7.3",
]

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I", "B", "W", "UP"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_unused_ignores = true
mypy_path = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=app --cov-report=xml --cov-report=term"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
